+++ src/opcodes/ldvxdt.go
@@ -2,7 +2,7 @@ package LDVXDT
 
 // LDVXDT Fx07 - Set Vx = delay timer value.
 func LDVXDT(c *cpu, opcode uint16) {
-	var x uint8 = uint8((opcode & 0x0F00) >> 8)
+	var x uint8 = (opcode & 0x0F00) >> 8
 
 	c.register[x] = c.systemTimer
 }
