+++ src/opcodes/rndvxb.go
@@ -7,8 +7,8 @@ import (
 
 // RNDVXB Cxkk - Set Vx = random byte AND kk
 func RNDVXB(c *cpu, opcode uint16) {
-	var x uint8 = uint8((opcode & 0x0F00) >> 8)
-	var kk uint8 = uint8((opcode & 0x00FF))
+	var x uint8 = (opcode & 0x0F00) >> 8
+	var kk uint8 = (opcode & 0x00FF)
 
 	rand.Seed(time.Now().UnixNano())
